% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe-svg-input.R
\name{globeSvgInput}
\alias{globeSvgInput}
\title{SVG Globe Input}
\usage{
globeSvgInput(inputId, ...)
}
\arguments{
\item{inputId}{The input and CSS identifier}

\item{...}{Child HTML elements}
}
\description{
SVG Globe Input
}
