% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3globe-package.R
\docType{package}
\name{r2d3globe-package}
\alias{r2d3globe}
\alias{r2d3globe-package}
\title{r2d3globe: Intractive Globes for 'Rmarkdown' and 'shiny' using 'D3'}
\description{
Provides inputs, outputs, and widgets to include interactive
  globes for 'Rmarkdown' and 'shiny' using 'D3'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/r2d3globe}
  \item Report bugs at \url{https://github.com/paleolimbot/r2d3globe/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
